aqueous: consolidate more aqueous specific stuff
